Circle refactor: sub-component for circle detail & edition OK
