+++ lib/frontend/src/reducer/index.js
@@ -222,10 +222,8 @@ const circle = (state = initial.circle, action) => {
   switch (action.type) {
     case 'EDIT_CIRCLE':
       return {
-        results: {
-          ...state.results,
-          is_in_edition: !state.results.is_in_edition,
-        },
+        ...state,
+        is_in_edition: !state.is_in_edition,
       }
     case 'TOGGLE_CIRCLE_ACCOUNT':
       return {
